+++ go/logger/logger.go
@@ -1,3 +1,23 @@
+//
+// Copyright (C) 2021 IBM Corporation.
+//
+// Authors:
+// Frederico Araujo <frederico.araujo@ibm.com>
+// Teryl Taylor <terylt@ibm.com>
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package logger implements logging utilities.
 package logger
 
 import (
@@ -18,15 +38,16 @@ const (
 	INFO
 	WARN
 	ERROR
+	HEALTH
 )
 
 func (d LogLevel) String() string {
-	return [...]string{"Trace", "Info", "Warn", "Error"}[d]
+	return [...]string{"Trace", "Info", "Warn", "Error", "Health"}[d]
 }
 
 // GetLogLevelFromValue returns LogLevel corresponding to string s (if not found, defaults to INFO).
 func GetLogLevelFromValue(s string) LogLevel {
-	m := map[string]LogLevel{"trace": TRACE, "info": INFO, "warn": WARN, "error": ERROR}
+	m := map[string]LogLevel{"trace": TRACE, "info": INFO, "warn": WARN, "error": ERROR, "health": HEALTH}
 	if l, ok := m[strings.ToLower(s)]; ok {
 		return l
 	}
@@ -35,29 +56,28 @@ func GetLogLevelFromValue(s string) LogLevel {
 
 // Loggers reflecting different log levels.
 var (
-	Trace *log.Logger
-	Info  *log.Logger
-	Warn  *log.Logger
-	Error *log.Logger
+	Trace  *log.Logger
+	Info   *log.Logger
+	Warn   *log.Logger
+	Error  *log.Logger
+	Health *log.Logger
 )
 
 // InitLoggers initialize utility loggers with default i/o streams.
 func InitLoggers(level LogLevel) {
 	switch level {
 	case TRACE:
-		initLoggers(os.Stdout, os.Stdout, os.Stdout, os.Stderr)
-		break
+		initLoggers(os.Stdout, os.Stdout, os.Stdout, os.Stderr, os.Stdout)
 	case INFO:
-		initLoggers(ioutil.Discard, os.Stdout, os.Stdout, os.Stderr)
-		break
+		initLoggers(ioutil.Discard, os.Stdout, os.Stdout, os.Stderr, os.Stdout)
 	case WARN:
-		initLoggers(ioutil.Discard, ioutil.Discard, os.Stdout, os.Stderr)
-		break
+		initLoggers(ioutil.Discard, ioutil.Discard, os.Stdout, os.Stderr, os.Stdout)
 	case ERROR:
-		initLoggers(ioutil.Discard, ioutil.Discard, ioutil.Discard, os.Stderr)
-		break
+		initLoggers(ioutil.Discard, ioutil.Discard, ioutil.Discard, os.Stderr, os.Stdout)
+	case HEALTH:
+		initLoggers(ioutil.Discard, ioutil.Discard, ioutil.Discard, os.Stderr, os.Stdout)
 	default:
-		initLoggers(ioutil.Discard, os.Stdout, os.Stdout, os.Stderr)
+		initLoggers(ioutil.Discard, os.Stdout, os.Stdout, os.Stderr, os.Stdout)
 	}
 }
 
@@ -65,7 +85,8 @@ func initLoggers(
 	traceHandle io.Writer,
 	infoHandle io.Writer,
 	warnHandle io.Writer,
-	errorHandle io.Writer) {
+	errorHandle io.Writer,
+	healthHandle io.Writer) {
 
 	Trace = log.New(traceHandle,
 		fmt.Sprintf("[%s] ", TRACE),
@@ -81,5 +102,9 @@ func initLoggers(
 
 	Error = log.New(errorHandle,
 		fmt.Sprintf("[%s] ", ERROR),
+		log.Ldate|log.Ltime|log.Lshortfile)
+
+	Health = log.New(healthHandle,
+		fmt.Sprintf("[%s] ", HEALTH),
 		log.Ldate|log.Ltime|log.Lshortfile)
 }