+++ py3/classes/actiontype/sysflow/type/__init__.py
@@ -5,7 +5,7 @@
 # Authors:
 # Frederico Araujo <frederico.araujo@ibm.com>
 # Teryl Taylor <terylt@ibm.com>
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -19,4 +19,5 @@
 # limitations under the License.
 #
 from ...schema_classes import SchemaClasses
+
 ActionType = SchemaClasses.sysflow.type.ActionTypeClass