+++ py3/classes/oid/__init__.py
@@ -5,7 +5,7 @@
 # Authors:
 # Frederico Araujo <frederico.araujo@ibm.com>
 # Teryl Taylor <terylt@ibm.com>
-# 
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -25,17 +25,18 @@ from avro.io import DatumReader
 
 
 class SpecificDatumReader(DatumReader):
-    SCHEMA_TYPES = {
-    }
+    SCHEMA_TYPES = {}
+
     def __init__(self, readers_schema=None, **kwargs):
         writers_schema = kwargs.pop("writers_schema", readers_schema)
         writers_schema = kwargs.pop("writer_schema", writers_schema)
         super(SpecificDatumReader, self).__init__(writers_schema, readers_schema, **kwargs)
+
     def read_record(self, writers_schema, readers_schema, decoder):
-        
+
         result = super(SpecificDatumReader, self).read_record(writers_schema, readers_schema, decoder)
-        
+
         if readers_schema.fullname in SpecificDatumReader.SCHEMA_TYPES:
             result = SpecificDatumReader.SCHEMA_TYPES[readers_schema.fullname](result)
-        
+
         return result