+++ py3/classes/sysflow/sysflow/flow/__init__.py
@@ -1,4 +1,5 @@
 from ...schema_classes import SchemaClasses
+
 FileFlow = SchemaClasses.sysflow.flow.FileFlowClass
 NetworkFlow = SchemaClasses.sysflow.flow.NetworkFlowClass
 ProcessFlow = SchemaClasses.sysflow.flow.ProcessFlowClass