+++ py3/classes/sysflow/sysflow/type/__init__.py
@@ -1,4 +1,5 @@
 from ...schema_classes import SchemaClasses
+
 ContainerType = SchemaClasses.sysflow.type.ContainerTypeClass
 OID = SchemaClasses.sysflow.type.OIDClass
 SFObjectState = SchemaClasses.sysflow.type.SFObjectStateClass